+++ src/olympia/blocklist/utils.py
@@ -1,3 +1,6 @@
+from django.utils.html import format_html, conditional_escape
+from django.utils.safestring import mark_safe
+
 from olympia import amo
 from olympia.activity import log_create
 
@@ -22,3 +25,44 @@ def block_activity_log_delete(obj, user):
     log_create(
         amo.LOG.BLOCKLIST_BLOCK_DELETED, obj.addon, obj.guid, obj,
         details={'guid': obj.guid}, user=user)
+
+
+def format_block_history(logs):
+    def format_html_join_kw(sep, format_string, kwargs_generator):
+        return mark_safe(conditional_escape(sep).join(
+            format_html(format_string, **kwargs)
+            for kwargs in kwargs_generator
+        ))
+
+    history_format_string = (
+        '<li>'
+        '{date}. {action} by {name}: {guid}{versions}. {legacy}'
+        '<ul><li>{reason}</li></ul>'
+        '</li>')
+    guid_url_format_string = '<a href="{url}">{text}</a>'
+    versions_format_string = ', versions {min} - {max}'
+
+    log_entries_gen = (
+        {'date': (
+            format_html(
+                guid_url_format_string,
+                url=log.details.get('url'),
+                text=log.created.date())
+            if log.details.get('url') else log.created.date()),
+         'action': amo.LOG_BY_ID[log.action].short,
+         'name': log.author_name,
+         'guid': log.details.get('guid'),
+         'versions': (
+            format_html(
+                versions_format_string, **{
+                    'min': log.details.get('min_version'),
+                    'max': log.details.get('max_version')})
+            if 'min_version' in log.details else ''),
+         'legacy': (
+            'Included in legacy blocklist.'
+            if log.details.get('include_in_legacy') else ''),
+         'reason': log.details.get('reason') or ''}
+        for log in logs)
+    return format_html(
+        '<ul>\n{}\n</ul>',
+        format_html_join_kw('\n', history_format_string, log_entries_gen))
